requirements.txt
pydgraph
pymongo
cassandra-driver
bcrypt
python-dotenv
